requirements.txt
tensorflow
nudenet
discord.py
utilsx
python-dotenv
